Restructure the directory to introduce src directory
