+++ protocol-specs/index.js
@@ -21,18 +21,23 @@ const fs = require('fs');
 const { join: pathJoin, extname } = require('path');
 const { execSync } = require('child_process');
 
-const generators = fs.readdirSync('./generators');
+const generators = fs.readdirSync('./src/generators');
 
 // eslint-disable-next-line no-restricted-syntax
 for (const aGenerator of generators) {
 	// if its a directory, there is a base_generator.js file as well
 	if (!extname(aGenerator)) {
-		const path = pathJoin(__dirname, './generators', aGenerator, 'index.js');
+		const path = pathJoin(
+			__dirname,
+			'./src/generators',
+			aGenerator,
+			'index.js',
+		);
 		console.log(`Executing generator '${aGenerator}' in path '${path}'`);
 		execSync(`node ${path}`);
 	}
 }
 console.log();
 console.log(
-	`All specs available in '${pathJoin(__dirname, '../generator_outputs/')}'`,
+	`All specs available in '${pathJoin(__dirname, '../../generator_outputs/')}'`,
 );
