+++ protocol-specs/src/generators/base_generator.js
@@ -24,7 +24,7 @@ const runGenerator = (generatorName, suiteCreators) => {
 
 		const dirPath = path.join(
 			__dirname,
-			`../generator_outputs/${suiteResult.runner}`,
+			`../../generator_outputs/${suiteResult.runner}`,
 		); // TODO: remove hard path and maybe expose it to a CLI
 
 		if (!fs.existsSync(dirPath)) {
