+++ protocol-specs/src/generators/bft_processing/11_delegates_partial_switch.csv
