+++ protocol-specs/src/generators/bft_processing/4_delegates_simple.csv
