+++ protocol-specs/src/generators/bft_processing/5_delegates_switched_completely.csv
