+++ protocol-specs/src/generators/bft_processing/index.js
