+++ protocol-specs/src/generators/block_processing_delegate/README.md
