+++ protocol-specs/src/generators/transaction_signing/README.md
