+++ protocol-specs/src/generators/transaction_signing/index.js
