+++ protocol-specs/src/utils/bft/index.js
