+++ protocol-specs/src/utils/dpos/index.js
